Internal: create a module for bootloader hints (#12)
As we need to implement hints for the bootloader, simple bootloader and
other functions, it is time to split the structs and constants into
their own modules inside a common `bootloader` module.
