+++ vm/src/hint_processor/builtin_hint_processor/bootloader/vars.rs
@@ -1,0 +1,14 @@
+/// Deserialized bootloader input.
+pub(crate) const BOOTLOADER_INPUT: &str = "bootloader_input";
+
+/// Saved state of the output builtin.
+pub(crate) const OUTPUT_BUILTIN_STATE: &str = "output_builtin_state";
+
+/// Deserialized simple bootloader input.
+pub(crate) const SIMPLE_BOOTLOADER_INPUT: &str = "simple_bootloader_input";
+
+/// Packed outputs.
+pub(crate) const PACKED_OUTPUTS: &str = "packed_outputs";
+
+/// Packed output for the current task.
+pub(crate) const PACKED_OUTPUT: &str = "packed_output";
