+++ vm/src/hint_processor/builtin_hint_processor/mod.rs
@@ -1,7 +1,7 @@
 pub mod bigint;
 pub mod blake2s_hash;
 pub mod blake2s_utils;
-pub mod bootloader_hints;
+mod bootloader;
 pub mod builtin_hint_processor_definition;
 pub mod cairo_keccak;
 pub mod dict_hint_utils;
